fix(security): address PR #285 review — auth timing, embedder SSRF/billing, attribution, scope honesty
Resolves reviewer blockers/defects the prior hardening commit left open:

- D9: constant_time_compare no longer leaks key length / allows DoS via an
  attacker-controlled dummy compare. It enforces equal length (key length is
  public) then uses subtle::ct_eq, and the inaccurate "does not leak length"
  comment is removed. All 26 auth tests pass (incl. the 256-multiple regression).
- D7: HttpEmbedder::is_available uses HEAD + a 30s cache instead of POSTing a
  billable embedding request on every call.
- D8 / #284: the embedder carries a self-contained insecure-remote-URL guard
  (opt out via SHODH_ALLOW_INSECURE_REMOTE_EMBEDDER). The duplicate
  is_insecure_remote_url / resolve_api_url_override added to integrations/mod.rs
  are reverted so PR #284 remains the sole owner of those symbols — no
  duplicate-definition merge conflict.
- B1 honesty: narrowed the "opaque at rest" claim. The memory_index CF stores
  plaintext lookup keys (tag/entity/date/…); this is now documented as a known
  gap in SECURITY.md rather than implied away.
- B3 / egress docs: key-rotation limitation and embedder egress posture
  documented in SECURITY.md.
- Attribution: Apache-2.0 contribution headers on encryption.rs and
  http_embedder.rs + a NOTICE file. Portll is the original author; the veld
  cross-reference is informational and no BUSL-1.1 terms attach.
